Fix cell's type (replace i8 with i32)

+++ src/core/cellular_automaton.rs
@@ -3,10 +3,10 @@ use crate::core::lookup_table::M1DLookupTable;
 
 #[allow(dead_code)]
 pub struct CA1D {
-    pub current_world: Box<Vec<i8>>,
-    pub next_world: Box<Vec<i8>>,
+    pub current_world: Box<Vec<i32>>,
+    pub next_world: Box<Vec<i32>>,
     pub lookup_table: M1DLookupTable,
-    pub num_states: u8,
+    pub num_states: i32,
     pub num_cells: usize,
 }
 
@@ -41,7 +41,7 @@ impl sdt::fmt::Display for CA1D {
 #[allow(dead_code)]
 impl CA1D {
     pub fn new(
-        num_states: u8,
+        num_states: i32,
         num_cells: usize,
         lookup_table: M1DLookupTable,
     ) -> CA1D {
@@ -56,15 +56,15 @@ impl CA1D {
         }
     }
 
-    pub fn get_current_world(&self) -> &Vec<i8> {
+    pub fn get_current_world(&self) -> &Vec<i32> {
         &self.current_world
     }
 
-    pub fn get_next_world(&self) -> &Vec<i8> {
+    pub fn get_next_world(&self) -> &Vec<i32> {
         &self.next_world
     }
 
-    pub fn get_num_states(&self) -> u8 {
+    pub fn get_num_states(&self) -> i32 {
         self.num_states
     }
 
@@ -76,19 +76,19 @@ impl CA1D {
         &self.lookup_table
     }
 
-    pub fn set_current_world(&mut self, current_world: Vec<i8>) {
+    pub fn set_current_world(&mut self, current_world: Vec<i32>) {
         self.current_world = Box::new(current_world);
     }
 
-    pub fn set_next_world(&mut self, next_world: Vec<i8>) {
+    pub fn set_next_world(&mut self, next_world: Vec<i32>) {
         self.next_world = Box::new(next_world);
     }
 
-    pub fn set_world_initialisation(&mut self, world_initialisation: Vec<i8>) {
+    pub fn set_world_initialisation(&mut self, world_initialisation: Vec<i32>) {
         self.current_world = Box::new(world_initialisation);
     }
 
-    pub fn set_num_states(&mut self, num_states: u8) {
+    pub fn set_num_states(&mut self, num_states: i32) {
         self.num_states = num_states;
     }
